01 - HTTP Methods
Congratulations! You finished the first class of our course. And what do we learn?

Changing with PUT: We have seen that, since we are able to add and read entries in our resource via HTTP, the next natural step is to allow users to be able to change these entries. For this, we use the PUT method in HTTP, passing the input ID to be changed in the URL and the entire body of the Journal object as JSON in the body of the request.

Removing with DELETE: Finally, it was necessary to remove an entry. We have seen that, in fact, this is one of the simplest operations for our case. It was enough to use the HTTP DELETE method, passing the input ID to be changed in the URL.

Confirmation Dialog: Because removal is a destructive operation, we note that it would be interesting to add a layer of protection against accidental cases. For this we created a generic confirmation Dialog (generic, because it can be used in other situations), and only when the user confirms the desire to perform this operation will we remove the entry.