02 - Authorization Flux
Class 2 finished successfully? So, let's remember what we learned:

Installing JSON Server Auth: We saw that, as it was, our server was not prepared to handle user authentication. To enable this functionality we have installed JSON Server Auth.

Understanding the Authorization Flow: We talked about that each API will have specificities in the way of authenticating and authorizing users, but that the system that uses User - Password - Token, or just Token Authentication, is very common and is what we will use with the JSON Server Auth.

Postman: We know Postman, a tool for executing and detailing requests and responses. It is very useful for us to do test operations quickly for our API, freely changing the information sent (such as headers and body) and carefully analyzing the received response.