03 - Authentication on Flutter
Another finished class and how about we review what we learned?

Importing Login Screen: We have seen that we will not always need to create a tool from scratch, often part of the code we are going to use is already ready and just import it into our code. This was exactly the case of our Login Screen, we received the code in Dart ready and we just need to import it into our file directory and make small settings.

Logging in via API: Soon in the sequence we needed to send the email and password information to the API, so that the server could try to authenticate with this information. For this, we capture this data on the screen we import and create a Service called auth_service. We did this to separate the requests for Journal from those related to authentication. Then just send the email and password information via POST, and receive the response from the server.

Registering via API: We noticed that, in order not to need to create a screen for registration, we could suggest the registration of a new user person if the email passed in the Login was invalid. To do this test, we checked the Status Code of the login response and, in case of failure, we checked if the body of the response contained the text we saw to be the error message for cases where the server did not find the email passed in the request in its database. In these cases, we open a confirmation Dialog suggesting the user to create a new account with this email and password.

Saving Locally: Finally, we know that soon on we will use the Token and user ID information to make our authenticated requests. Therefore, we have already done the work of saving persistently, using Shared Preferences, this information in the device memory.